fix(project): support windows in worker graceful stop
Signal(SIGTERM) and Signal(0) return errors on windows, breaking
worker --graceful-stop-limit. Split the graceful-stop logic into
platform files: unix keeps SIGTERM-then-kill, windows kills directly.
