+++ cmd/project/project_worker.go
@@ -104,21 +104,7 @@ var projectWorkerCmd = &cobra.Command{
 					cmd.Env = append(os.Environ(), fmt.Sprintf("MESSENGER_CONSUMER_NAME=%s-%d", baseName, index))
 					cmd.WaitDelay = time.Second
 					cmd.Cancel = func() error {
-						if gracefulStopLimit > 0 {
-							if err := cmd.Process.Signal(syscall.SIGTERM); err != nil {
-								return err
-							}
-
-							now := time.Now()
-
-							for time.Since(now) < time.Second*time.Duration(gracefulStopLimit) {
-								if isProcessStopped(cmd.Process) {
-									return os.ErrProcessDone
-								}
-								time.Sleep(time.Millisecond * 250)
-							}
-						}
-						return cmd.Process.Kill()
+						return gracefulStop(cmd, gracefulStopLimit)
 					}
 
 					if err := cmd.Run(); err != nil {
@@ -156,8 +142,4 @@ func cancelOnTermination(ctx context.Context, cancel context.CancelFunc) {
 		logging.FromContext(ctx).Infof("received signal %v\n", sig.String())
 		cancel()
 	}()
-}
-
-func isProcessStopped(p *os.Process) bool {
-	return errors.Is(p.Signal(syscall.Signal(0)), os.ErrProcessDone)
 }
